feat: 내 정보 조회 기능 완성

+++ src/main/java/Qpick/server/domain/product/domain/repository/ProductManualRepository.java
@@ -3,5 +3,5 @@ package Qpick.server.domain.product.domain.repository;
 import Qpick.server.domain.product.domain.entity.ProductManual;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductManualRepository extends JpaRepository<ProductManual,Integer> {
+public interface ProductManualRepository extends JpaRepository<ProductManual, Long> {
 }
